+++ packages/editor/src/lib/config/defaultShapes.ts
@@ -1,26 +1,21 @@
-import { TLGroupShape, TLImageShape, TLVideoShape } from '@tldraw/tlschema'
+import { TLImageShape, TLVideoShape } from '@tldraw/tlschema'
 import { arrowShape } from '../editor/shapes/ArrowShape/arrowShape'
 import { bookmarkShape } from '../editor/shapes/BookmarkShape/bookmarkShape'
 import { drawShape } from '../editor/shapes/DrawShape/drawShape'
 import { embedShape } from '../editor/shapes/EmbedShape/embedShape'
 import { frameShape } from '../editor/shapes/FrameShape/frameShape'
 import { geoShape } from '../editor/shapes/GeoShape/geoShape'
+import { groupShape } from '../editor/shapes/GroupShape/groupShape'
 import { highlightShape } from '../editor/shapes/HighlightShape/highlightShape'
 import { lineShape } from '../editor/shapes/LineShape/lineShape'
 import { noteShape } from '../editor/shapes/NoteShape/noteShape'
 import { textShape } from '../editor/shapes/TextShape/textShape'
-import { GroupShapeUtil } from '../editor/shapeutils/GroupShapeUtil/GroupShapeUtil'
 import { ImageShapeUtil } from '../editor/shapeutils/ImageShapeUtil/ImageShapeUtil'
 import { VideoShapeUtil } from '../editor/shapeutils/VideoShapeUtil/VideoShapeUtil'
 import { TLShapeInfo, createShape } from './createShape'
 
 /** @public */
 export const coreShapes: TLShapeInfo<any>[] = [
-	// created by grouping interactions, probably the corest core shape that we have
-	createShape<TLGroupShape>('group', {
-		util: GroupShapeUtil,
-	}),
-	// created by copy and paste / file drop
 	createShape<TLImageShape>('image', {
 		util: ImageShapeUtil,
 	}),
@@ -42,4 +37,5 @@ export const defaultShapes: TLShapeInfo<any>[] = [
 	textShape,
 	bookmarkShape,
 	embedShape,
+	groupShape,
 ]
