+++ packages/editor/src/lib/editor/shapes/GroupShape/groupShape.ts
@@ -1,0 +1,12 @@
+import { createShape } from '../../../config/createShape'
+import { GroupShapeUtil } from './GroupShapeUtil/GroupShapeUtil'
+import { groupShapeMigrations } from './groupShapeMigrations'
+import { TLGroupShape } from './groupShapeTypes'
+import { groupShapeValidator } from './groupShapeValidator'
+
+/** @public */
+export const groupShape = createShape<TLGroupShape>('group', {
+	util: GroupShapeUtil,
+	migrations: groupShapeMigrations,
+	validator: groupShapeValidator,
+})
