+++ packages/editor/src/lib/editor/shapes/GroupShape/groupShapeMigrations.ts
@@ -1,0 +1,4 @@
+import { defineMigrations } from '@tldraw/store'
+
+/** @internal */
+export const groupShapeMigrations = defineMigrations({})
