+++ packages/editor/src/lib/editor/shapes/GroupShape/groupShapeTypes.ts
@@ -1,0 +1,7 @@
+import { TLBaseShape } from '@tldraw/tlschema'
+
+/** @public */
+export type TLGroupShapeProps = { [key in never]: undefined }
+
+/** @public */
+export type TLGroupShape = TLBaseShape<'group', TLGroupShapeProps>
