+++ packages/editor/src/lib/editor/shapes/GroupShape/groupShapeValidator.ts
@@ -1,0 +1,9 @@
+import { createShapeValidator } from '@tldraw/tlschema'
+import { T } from '@tldraw/validate'
+import { TLGroupShape } from './groupShapeTypes'
+
+/** @internal */
+export const groupShapeValidator: T.Validator<TLGroupShape> = createShapeValidator(
+	'group',
+	T.object({})
+)
