+++ packages/editor/src/lib/test/tools/groups.test.ts
@@ -1,15 +1,16 @@
 import { sortByIndex } from '@tldraw/indices'
 import { approximately, Box2d, VecLike } from '@tldraw/primitives'
-import { createShapeId, TLGroupShape, TLShape, TLShapeId, TLShapePartial } from '@tldraw/tlschema'
+import { createShapeId, TLShape, TLShapeId, TLShapePartial } from '@tldraw/tlschema'
 import { assert, compact } from '@tldraw/utils'
 import { ArrowShapeTool } from '../../editor/shapes/ArrowShape/ArrowShapeTool/ArrowShapeTool'
 import { TLArrowShape } from '../../editor/shapes/ArrowShape/arrowShapeTypes'
 import { ArrowShapeUtil } from '../../editor/shapes/ArrowShape/ArrowShapeUtil/ArrowShapeUtil'
 import { DrawShapeTool } from '../../editor/shapes/DrawShape/DrawShapeTool/DrawShapeTool'
+import { TLGroupShape } from '../../editor/shapes/GroupShape/groupShapeTypes'
+import { GroupShapeUtil } from '../../editor/shapes/GroupShape/GroupShapeUtil/GroupShapeUtil'
 import { LineShapeTool } from '../../editor/shapes/LineShape/LineShapeTool/LineShapeTool'
 import { TLLineShape } from '../../editor/shapes/LineShape/lineShapeTypes'
 import { NoteShapeTool } from '../../editor/shapes/NoteShape/NoteShapeTool/NoteShapeTool'
-import { GroupShapeUtil } from '../../editor/shapeutils/GroupShapeUtil/GroupShapeUtil'
 import { EraserTool } from '../../editor/tools/EraserTool/EraserTool'
 import { TestEditor } from '../TestEditor'
